chore: refine parser and expand tests
